+++ src/components/ArrowsInline/index.js
@@ -1,18 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames';
 
 /**
  * @param {Object} props - Properties passed to component
  * @returns {ReactElement}
  */
-const ArrowsInline = (props) => (
-  <span
-    data-oui-component={ true }
-    className={ 'oui-arrow-inline--' + props.direction }
-    aria-hidden="true"
-    data-test-section={ props.testSection }>
-  </span>
-);
+
+const ArrowsInline = (props) => {
+  let classes = classNames({
+    'highlight-react--oui': localStorage.getItem('show_ouireact') === 'true',
+    [`oui-arrow-inline--${props.direction}`]: props.direction,
+  });
+
+  return (
+    <span
+      data-oui-component={ true }
+      className={ classes }
+      aria-hidden="true"
+      data-test-section={ props.testSection }>
+    </span>
+  );
+};
 
 ArrowsInline.propTypes = {
   /** Direction the arrow points */