+++ src/components/BlockList/index.js
@@ -16,6 +16,7 @@ const BlockList = (props) => {
     'background--white': true,
     'overflow-x--auto': true,
     'border--all': props.hasBorder,
+    'highlight-react--oui': localStorage.getItem('show_ouireact') === 'true',
   });
 
   return (