+++ src/components/Button/index.js
@@ -13,7 +13,9 @@ const Button = (props) => {
     [`oui-button--${props.size}`]: props.size,
     [`oui-button--${props.width}`]: props.width,
     ['is-active']: props.isActive,
+    'highlight-react--oui': localStorage.getItem('show_ouireact') === 'true',
   });
+
   return (
     <button
       data-oui-component={ true }